+++ packages/pangea-sdk/PangeaCyber.Tests/ITAIGuardTest.cs
@@ -14,16 +14,30 @@ public class ITAIGuardTest
     {
         var response = await client.GuardText(new TextGuardRequest("hello world") { Recipe = "pangea_prompt_guard" });
         Assert.True(response.IsOK);
-        Assert.NotNull(response.Result.Prompt);
-        Assert.False(response.Result.Detectors.MaliciousEntity?.Detected);
+        Assert.NotNull(response.Result.PromptText);
+        if (response.Result.Detectors.MaliciousEntity != null)
+        {
+            Assert.False(response.Result.Detectors.MaliciousEntity.Detected);
+        }
         Assert.False(response.Result.Detectors.PiiEntity?.Detected);
         Assert.False(response.Result.Detectors.PromptInjection?.Detected);
 
         response = await client.GuardText(new TextGuardRequest("security@pangea.cloud") { Recipe = "pangea_prompt_guard" });
         Assert.True(response.IsOK);
-        Assert.NotNull(response.Result.Prompt);
+        Assert.NotNull(response.Result.PromptText);
         Assert.NotNull(response.Result.Detectors.PiiEntity);
         Assert.NotNull(response.Result.Detectors.PiiEntity.Data);
         Assert.Single(response.Result.Detectors.PiiEntity.Data.Entities);
+    }
+
+    [Fact]
+    public async Task TestGuardMessages()
+    {
+        var response = await client.GuardText(new MessagesGuardRequest<IEnumerable<IDictionary<string, string>>>(new[]
+        {
+            new Dictionary<string, string> { { "role", "user" }, { "content", "what was pangea?" } }
+        }));
+        Assert.True(response.IsOK);
+        Assert.NotNull(response.Result.PromptMessages);
     }
 }